rename files and add readme for llama auto_parallel (#7944)
* rename files and add readme for llama auto_parallel

* rename files and add readme for llama auto_parallel

* fix ci
