+++ llm/llama/auto_parallel/run_pretrain_auto.sh
@@ -17,9 +17,6 @@
 set -x
 unset CUDA_VISIBLE_DEVICES
 
-export FLAGS_call_stack_level=3
-export FLAGS_use_cuda_managed_memory=true
-
 task_name="llama_auto_dp2mp2pp2"
 rm -rf output/$task_name/
 rm -rf "output/$task_name""_log"
@@ -29,16 +26,19 @@ export PYTHONPATH=../../../:$PYTHONPATH
 #ulimit -c unlimited
 #export GLOG_v=10
 
-rm -rf log_auto
+# export FLAGS_call_stack_level=3
+# export FLAGS_use_cuda_managed_memory=true
 
-export FLAGS_embedding_deterministic=1        
-export FLAGS_cudnn_deterministic=1
-export NVIDIA_TF32_OVERRIDE=0
+# export FLAGS_embedding_deterministic=1        
+# export FLAGS_cudnn_deterministic=1
+# export NVIDIA_TF32_OVERRIDE=0
 
-python3.8 -u  -m paddle.distributed.launch \
+to_static=0  # 是否开启动转静训练
+
+python -u  -m paddle.distributed.launch \
     --gpus "0,1,2,3,4,5,6,7" \
     --log_dir "auto_3d" \
-    run_pretrain_3D_auto.py \
+    run_pretrain_auto.py \
     --model_type "llama" \
     --model_name_or_path "facebook/llama-7b" \
     --tokenizer_name_or_path "facebook/llama-7b" \
@@ -76,3 +76,4 @@ python3.8 -u  -m paddle.distributed.launch \
     --data_impl "mmap" \
     --enable_auto_parallel 1 \
     --max_grad_norm 1.0 \
+    --to_static $to_static \
