+++ llm/llama/auto_parallel/run_pretrain_auto_static_sp.sh
@@ -17,7 +17,7 @@
 set -x
 unset CUDA_VISIBLE_DEVICES
 
-task_name="llama_auto_dp2sharding2mp2pp2_vpp2"
+task_name="llama_auto_dp2mp2pp2_vpp2_sp"
 # rm -rf output/$task_name/  # ckpt is saved in 'output/''
 rm -rf "output/$task_name""_log"
 
@@ -27,7 +27,7 @@ export PYTHONPATH=../../../:$PYTHONPATH
 python -u -m paddle.distributed.launch \
     --gpus "0,1,2,3,4,5,6,7" \
     --log_dir "output/$task_name""_log" \
-    run_pretrain_auto.py \
+    run_pretrain_auto_static.py \
     --model_type "llama" \
     --model_name_or_path "facebook/llama-7b" \
     --tokenizer_name_or_path "facebook/llama-7b" \
@@ -47,7 +47,7 @@ python -u -m paddle.distributed.launch \
     --pipeline_parallel_degree 2 \
     --virtual_pp_degree 2 \
     --pipeline_schedule_mode "VPP" \
-    --sharding_parallel_degree 2 \
+    --sharding_parallel_degree 1 \
     --sharding "stage2" \
     --learning_rate 0.0001 \
     --min_learning_rate 0.00001 \
@@ -68,6 +68,7 @@ python -u -m paddle.distributed.launch \
     --do_eval \
     --device "gpu" \
     --data_impl "mmap" \
-    --enable_auto_parallel 1
+    --enable_auto_parallel 1 \
+    --sequence_parallel true \
 
     # --resume_from_checkpoint "output/llama_auto_serial/checkpoint-2" \
