+++ llm/llama/auto_parallel/run_pretrain_hand.py
