+++ paddlenlp/transformers/llama/__init__.py
@@ -14,7 +14,7 @@
 
 from .configuration import *
 from .modeling import *
-from .modeling_3D_auto import *
 from .modeling_auto import *
+from .modeling_auto_static import *
 from .modeling_pp import *
 from .tokenizer import *
